# Improvement of sub-categories count