--- a/src/Filters/Block.php
+++ b/src/Filters/Block.php
@@ -1017,7 +1017,7 @@ class Block
                 continue;
             }
 
-            if ($categoryCount && empty($categoriesId) === false && in_array($idCategory, $categoriesId) === false) {
+            if ($categoryCount && in_array($idCategory, $categoriesId) === false) {
                 continue;
             }