--- a/src/Filters/Block.php
+++ b/src/Filters/Block.php
@@ -1005,7 +1005,7 @@ class Block
             $query->where('nleft > ' . (int) $parent->nleft);
             $query->where('nright < ' . (int) $parent->nright);
             $resultCategories = \Db::getInstance()->executeS($query);
-            foreach($resultCategories as $oneCategory) {
+            foreach ($resultCategories as $oneCategory) {
                 $categoriesId[] = $oneCategory['id_category'];
             }
         }