--- a/src/Filters/Block.php
+++ b/src/Filters/Block.php
@@ -989,27 +989,22 @@ class Block
             $categories[$value['id_category']] = $value;
         }
 
-        $categoryCount = true;
-        if (true === defined('_PS_FACETED_NEWCOUNT_') && null !== _PS_FACETED_NEWCOUNT_) {
-            $categoryCount = (bool) _PS_FACETED_NEWCOUNT_;
-        }
-        $results = $filteredSearchAdapter->valueCount('id_category', $categoryCount);
-
+        $results = $filteredSearchAdapter->valueCount('id_category');
+        // To filter results, instead to make a complex query, we will filter them in the code after
+        // So we retrieve here all categories linked to the parent one according to configuration (only IDs)
         $categoriesId = [];
-        if ($categoryCount === true) {
-            $query = new \DbQuery();
-            $query->select('id_category');
-            $query->from('category');
-            $depth = (int) Configuration::get('PS_LAYERED_FILTER_CATEGORY_DEPTH', null, null, null, 1);
-            if ($depth > 0) {
-                $query->where('level_depth <= ' . ((int) $parent->level_depth + $depth));
-            }
-            $query->where('nleft > ' . (int) $parent->nleft);
-            $query->where('nright < ' . (int) $parent->nright);
-            $resultCategories = \Db::getInstance()->executeS($query);
-            foreach ($resultCategories as $oneCategory) {
-                $categoriesId[] = $oneCategory['id_category'];
-            }
+        $query = new \DbQuery();
+        $query->select('id_category');
+        $query->from('category');
+        $depth = (int) Configuration::get('PS_LAYERED_FILTER_CATEGORY_DEPTH', null, null, null, 1);
+        if ($depth > 0) {
+            $query->where('level_depth <= ' . ((int) $parent->level_depth + $depth));
+        }
+        $query->where('nleft > ' . (int) $parent->nleft);
+        $query->where('nright < ' . (int) $parent->nright);
+        $resultCategories = \Db::getInstance()->executeS($query);
+        foreach ($resultCategories as $oneCategory) {
+            $categoriesId[] = $oneCategory['id_category'];
         }
 
         foreach ($results as $key => $values) {
@@ -1020,7 +1015,8 @@ class Block
                 continue;
             }
 
-            if ($categoryCount && in_array($idCategory, $categoriesId) === false) {
+            // If category count is not in the list of categories linked to the parent one, we skip it
+            if (in_array($idCategory, $categoriesId) === false) {
                 continue;
             }