<!-----------------------------------------------------------------------------
Thank you for contributing to the PrestaShop project! 

Please take the time to edit the "Answers" rows below with the necessary information.

Check out our contribution guidelines to find out how to complete it:
https://devdocs.prestashop.com/1.7/contribute/contribution-guidelines/#pull-requests
------------------------------------------------------------------------------>

| Questions     | Answers
| ------------- | -------------------------------------------------------
| Description  | When category count is enabled, the SQL query to retrieve the number of products per sub-category on a category page was particularly slow, and could slow down or even bring the MySQL server to its knees. This enhancement splits the process into two queries (one to retrieve the category list) and the second to output the product count. Considerable performance gains (12x to 100x faster depending number of products / categories).
| Type         | improvement 
| BC breaks    | no
| Deprecations | no
| Fixed ticket | No issue opened
| How to test  | First of all, we need to display one category with filters and sub-categories count enabled (PS_LAYERED_FULL_TREE to 1 and PS_LAYERED_SHOW_QTIES to 1). Then the category must have one model with sub categories filters enabled (layered_selection_subcategories). It's better to test without cache of blocks filters enabled and without cache to be abble to see performance before / after. You can add the constant _PS_FACETED_NEWCOUNT_ in your defines.inc.php (or custom) to false to get previous count and compare before / after.

Just a short explaination of diagnostic and benchmark done, with ps profiling tool.
I noticed that the main slow query was pointing the categories count. 

---------------------

#### First case - Shop with 200k+ products - 415 categories - Display of category with 5500 products

**Before**
![image](https://github.com/user-attachments/assets/3e445779-3d28-4967-a244-588d29a27fef)

**After**
![image](https://github.com/user-attachments/assets/5d414a56-b00c-4d53-8230-e950ad8a3f4d)

As we can see we have following result : 

- **Before** - **11610 ms** with 253980 results - 2195 queries
- **After** - **931 ms** with 3344771556 results - 2196 queries (**12x faster**)
We have one more query to get the categories concerned by the filters after, and then count is reduced
and we have the same results in output.

---------------------

#### Second case - Shop with 5k+ products - 289 categories - Display of category with 700 products

**Before**
![image](https://github.com/user-attachments/assets/1b2b21b5-0717-4598-9ef4-be7b9e7b3728)

**After**
![image](https://github.com/user-attachments/assets/91fef507-667e-4730-a384-176d84eecaad)

As we can see we have following result : 

- **Before** - **3664 ms** with 28896 results - 2009 queries
- **After** - **27.5 ms** with 326163600 results - 2010 queries (**133x faster**)
Same thing one more query to get categories concerned instead of add it in request for count.

---------------------

To be sure that output was the same i had some json exports before after but feel free to make the
same tests and point to me any issue about this and how to reproduce.

There is some conditions remaining to allow us to test ([_PS_FACETED_NEWCOUNT_](https://github.com/202ecommerce/ps_facetedsearch/blob/dev/src/Filters/Block.php#L992)) i can quickly remove
this when all is approved.

<!-- Click the form's "Preview" button to make sure the table is functional in GitHub. Thank you! -->