@ps-jarvis 2025-04-25
<!-- PR_WELCOME -->
Hello @kgleizes!

This is your first pull request on [ps_facetedsearch](https://github.com/PrestaShop/ps_facetedsearch/) repository of the PrestaShop project.

Thank you, and welcome to this Open Source community!