@kpodemski 2026-08-29
> [!NOTE]
> Semi-automated review. The tests below were run by Claude Code on a real shop. I checked every result before posting. A human is reading and answering here, so please reply as usual.

Hi @kgleizes, sorry for the long wait on this one. I rebased your branch on current `dev` and tested it. Two things to report, and one question for you.

## 1. The group access issue @PrestaEdit raised is real

I reproduced it. I removed the Visitor group from a sub-category, then viewed the parent category as a guest:

| build | Categories filter block |
| --- | --- |
| `dev` | Home Accessories (8) |
| this PR | Home Accessories (8), **Stationery (3)** |
| PR + fix | Home Accessories (8) |

A category the visitor is not allowed to see appears in the filter block, with its product count. I confirmed it by logging the generated SQL: after the stripping, the WHERE of the count query is reduced to `c.level_depth<=3`, so the `id_group` condition is gone and nothing puts it back.

The fix is to add the group restriction to your new `$categoriesId` query. I also added `active = 1` and a `category_shop` join for multistore, plus a comment about the ordering between the `unset` calls and the string matching in `computeWhereConditions`.

Two smaller notes on @PrestaEdit's review, after testing:

- The missing `active = 1` does not currently show up, because `getAllCategoriesName()` is called with `$active = true` and the loop skips anything outside that list. Still worth adding, but it is not a live bug.
- Multistore I could not verify, single shop install here.

## 2. I could not reproduce the performance gain

This is the part I would like your help with. I tested on 100 000 products, 400 000 combinations, MariaDB 11.4, with `PS_LAYERED_FULL_TREE=1`, `PS_LAYERED_SHOW_QTIES=1` and the block cache off. I grew the category tree three times to get closer to your shop:

| categories | sub-categories under the browsed one | count query, `dev` | count query, PR |
| --- | --- | --- | --- |
| 34 | 0 | 0.545s | 0.541s |
| 418 | 16 | 0.162s | 0.163s |
| 2 434 | 100 | 0.166s | 0.167s |

Page time did not move either, against controls that stayed flat. I also tried `PS_LAYERED_FILTER_CATEGORY_DEPTH = 0`, same result.

Here the sub-category count is about 0.17s of a 4.4s page, so even removing it entirely could not produce 12x. Your "before" was 11.6s for that one query, which is 70 times slower than what I measure on a much larger tree. That gap looks like a query plan difference rather than a data size difference, so I am fairly sure your shop hits something this environment does not.

## Could you share a bit about your setup?

- MySQL or MariaDB, and which version?
- `EXPLAIN` of the slow count query on your shop, before your change
- Roughly how many products, and how deep the category tree is
- Value of `PS_LAYERED_FILTER_CATEGORY_DEPTH`

With the `EXPLAIN` we can find what the optimizer does differently, and then decide whether the split is the right fix or whether an index or a smaller change solves it. The environment I used is public if you want to compare: https://github.com/kpodemski/prestashop-performance-tester

To be clear, I am not saying your numbers are wrong. I just cannot reproduce them yet, and I would rather understand why before merging a change to this query.